moved titlebar to be below OS default title bar for more consistency through different OSs
