+++ src/renderer/src/components/TitleBar/TitleBar.module.css
@@ -1,33 +1,28 @@
 .title-bar {
-  position: sticky;
   display: flex;
-  justify-content: space-between;
+  justify-content: center;
   align-items: center;
-  width: 100%;
+  max-width: 100%;
   height: 30px;
   top: 0;
-  border-width: 1px 1px 0 1px;
-  border-style: solid;
-  border-color: #b8b8b8;
-  background-color: #f3f3f3;
-  -webkit-user-select: none;
   -webkit-app-region: drag;
+  padding: 0px 15px;
 }
 
-.inner {
+.controls {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  max-width: 1400px;
+  flex-grow: 1;
+}
+
+.left-controls {
   display: flex;
   flex-direction: row;
   justify-content: center;
   align-items: center;
   gap: 20px;
-  height: 100%;
-}
-
-.sub-controls {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  -webkit-app-region: no-drag;
   height: 100%;
 }
 
@@ -66,12 +61,13 @@
   color: grey;
 }
 
-.main-controls-box {
+.right-controls {
   display: flex;
   flex-direction: row;
   justify-content: center;
   align-items: center;
   -webkit-app-region: no-drag;
+  gap: 7px;
 }
 
 .main-controls {
@@ -100,12 +96,12 @@
   background-color: #e9e9e9;
 }
 
-.theme-icon {
+.mode-icon {
   width: 18px;
   height: 18px;
 }
 
-.dark {
+.dark-mode {
   @mixin dark {
     display: none;
   }
@@ -115,7 +111,7 @@
   }
 }
 
-.light {
+.light-mode {
   @mixin light {
     display: none;
   }
@@ -123,4 +119,15 @@
   @mixin dark {
     display: block;
   }
+}
+
+.dark {
+  background-color: #323232;
+  border-bottom: 1px solid #424242;
+}
+
+.light {
+  background-color: rgb(255, 255, 255);
+  border-bottom: 1px solid #dbdada;
+  box-shadow: 0 0.5px 3px 0 rgba(208, 208, 208, 0.2);
 }
